+++ color.h
@@ -18,7 +18,7 @@ using color_t = unsigned char;
  * @param arg The input integer.
  * @return A color_t with the same value.
  */
-color_t operator"" _c(unsigned long long arg) noexcept;
+color_t operator"" _c(uint64_t arg) noexcept;
 
 /**
  * Cast to underlying type.
